feat: activer mode test pour rnn
